Upgraded to hadoop 0.14.4; Got regressions passing on HDFS
git-svn-id: https://hypertable.googlecode.com/svn/trunk@476 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Hypertable/RangeServer/CellStoreTrailerV0.cc
@@ -65,7 +65,7 @@ void CellStoreTrailerV0::serialize(uint8_t *buf) {
   Serialization::encode_short(&buf, compression_type);
   Serialization::encode_int(&buf, (uint32_t)compression_ratio);
   Serialization::encode_short(&buf, version);
-  assert((buf-base) == CellStoreTrailerV0::size());
+  assert((buf-base) == (int)CellStoreTrailerV0::size());
 }
